Rename modules
Changed module names to be cleaner and more accurate to their purpose.
