+++ src/discard/globals.py
